Added a betting endpoint that calculates winnings for a set amount of games. Once predicted games home model has data, the endpoint will work

+++ BusinessLogic/Betting/IBettingCalculator.cs
@@ -5,7 +5,7 @@ namespace BusinessLogic.Betting
 {
     public interface IBettingCalculator
     {
-        Task<IEnumerable<ModelBetResult>> CalculateBetOutcomes(int year, int numberOfGames, decimal betAmount);
+        Task<IEnumerable<ModelBetResult>> CalculateBetOutcomes(int year, int numberOfGames, double betAmount);
     }
 }
